move all components and section in their corresponding folder to clean architecture and replace favicon
