+++ frontend/app/[locale]/calculator/_components/calculator-select-item.tsx
