+++ frontend/app/[locale]/calculator/_section/calculator-hero-section.tsx
