+++ frontend/app/[locale]/calculator/_section/calculator-link-section.tsx
