+++ frontend/app/[locale]/calculator/_section/calculator-selector-section.tsx
@@ -1,5 +1,5 @@
 import { getScopedI18n } from '@/locales/server';
-import CalculatorSelect from "@/app/[locale]/ui/calculator-select-item";
+import CalculatorSelect from "@/app/[locale]/calculator/_components/calculator-select-item";
 
 export default async function SelectorSection() {
   const scopedT = await getScopedI18n("calculatorPage");
