+++ frontend/app/[locale]/calculator/page.tsx
@@ -1,6 +1,6 @@
-import CalculatorHeroSection from './calculator-hero-section';
-import CalculatorLinkSection from './calculator-link-section';
-import SelectorSection from './calculator-selector-section';
+import CalculatorHeroSection from './_section/calculator-hero-section';
+import CalculatorLinkSection from './_section/calculator-link-section';
+import SelectorSection from './_section/calculator-selector-section';
 
 
 export default async function Calculator() {
