+++ frontend/app/[locale]/home/_section/WFI-article-section.tsx
