+++ frontend/app/[locale]/home/_section/book-announcement-section.tsx
