+++ frontend/app/[locale]/home/_section/hero-section.tsx
