+++ frontend/app/[locale]/methodology/_components/list-of-affition.tsx
