+++ frontend/app/[locale]/methodology/_components/suffering-synthesis-duration-table.tsx
