+++ frontend/app/[locale]/methodology/_section/introduction-section.tsx
