+++ frontend/app/[locale]/methodology/page.tsx
@@ -1,18 +1,15 @@
-import SufferingQuantificationSteps from './suffering-quantification-steps';
-import IntroductionSection from './introduction-section';
-import MethodDetailsSection from './method-details-section';
-import KeyResultsSection from './key-results-section';
+import SufferingQuantificationStepsSection from './_section/suffering-quantification-steps-section';
+import IntroductionSection from './_section/introduction-section';
+import MethodDetailsSection from './_section/method-details-section';
+import KeyResultsSection from './_section/key-results-section';
 
-import GoFurtherSection from '@/app/[locale]/ui/general/home-page/go-further-section';
-
+import GoFurtherSection from '@/app/[locale]/ui/_section/go-further-section';
 
 export default async function MethodologyPage() {
-
-
   return (
     <>
       <IntroductionSection />
-      <SufferingQuantificationSteps />
+      <SufferingQuantificationStepsSection />
       <MethodDetailsSection />
       <KeyResultsSection />
       <GoFurtherSection />
