--- frontend/app/[locale]/off/knowledge_panel.css
@@ -1,26 +1,0 @@
-.panel-content a {
-  color: #f97316; /* Orange-500 de Tailwind */
-  text-decoration: underline;
-}
-
-.panel-content a:hover {
-  color: #ea580c; /* Orange-600 de Tailwind */
-}
-
-.panel-content ul {
-  list-style-type: disc;
-  padding-left: 1.5rem;
-  margin: 0.5rem 0;
-}
-
-.panel-content li {
-  margin-bottom: 0.25rem;
-}
-
-.product-image {
-  width: auto;
-  height: auto;
-  max-width: 400px;
-  max-height: 300px;
-  object-fit: contain;
-}
