+++ frontend/app/[locale]/ui/_components/BoltIcon.tsx
