+++ frontend/app/[locale]/ui/_components/heroBanner.tsx
