+++ frontend/app/[locale]/ui/_components/logo-walfare-footprint.tsx
