+++ frontend/app/[locale]/ui/_components/suffering-stages-description.tsx
