+++ frontend/app/[locale]/ui/_section/link-section.tsx
