+++ frontend/app/[locale]/ui/_section/navbar.tsx
@@ -1,13 +1,13 @@
 import { getI18n } from '@/locales/server';
 
-import NavLinks from '@/app/[locale]/ui/general/elements/nav-links';
-import { LocaleSelect } from '../localselect';
-import Logo from './home-page/elements/logo-walfare-footprint';
+import NavLinks from '@/app/[locale]/ui/_components/nav-links';
+import { LocaleSelect } from '../_components/localselect';
+import Logo from '../_components/logo-walfare-footprint';
 
 export default async function Navbar() {
   const t = await getI18n();
 
-    const links= [
+  const links = [
     { name: t('NavBarLink.methodology'), href: '/methodology' },
     { name: t('NavBarLink.calculator'), href: '/calculator' },
     { name: t('NavBarLink.about'), href: '/about' },
@@ -19,17 +19,17 @@ export default async function Navbar() {
         <div className="flex flex-col md:flex-row md:items-center md:justify-between w-full gap-4">
           <div className="relative flex justify-between md:justify-start items-center w-full">
             <div className="flex justify-start items-center">
-              <Logo/>
+              <Logo />
             </div>
             <div className="grow mr-20 hidden md:block">
-              <NavLinks links={links}/>
+              <NavLinks links={links} />
             </div>
             <div className="absolute right-0">
-              <LocaleSelect/>
+              <LocaleSelect />
             </div>
           </div>
           <div className="grow block md:hidden">
-            <NavLinks links={links}/>
+            <NavLinks links={links} />
           </div>
         </div>
       </header>
