+++ frontend/app/[locale]/ui/_section/section-heading.tsx
